refactor(types): convert types.py to package structure for better organization
- Split monolithic apiconfig/types.py (274 lines) into logical submodules
- Create apiconfig/types/ package with categorized type definitions:
  * json.py - JSON data structures and processing types
  * http.py - HTTP protocols, methods, and request/response types
  * config.py - Configuration management types
  * auth.py - Authentication, token, and storage strategy types
  * extensions.py - Logging and hook extension types
- Maintain complete backward compatibility via __init__.py re-exports
- Preserve all existing public API imports without breaking changes
- Add comprehensive __all__ definitions for clear module interfaces

This refactoring improves code maintainability and makes type definitions
easier to locate while ensuring zero impact on existing functionality.

Verified: All 987 tests pass, mypy type checking successful, flake8 compliant
